Move common query function to common parent
