+++ ocd/packages/desktop/package.json
@@ -109,7 +109,6 @@
     },
     "linux": {
       "icon": "build/assets/icon.png",
-      "executableName": "${productName}-${version}.${ext}",
       "target": [
         "AppImage",
         "deb",
